requirements.txt
nvdlib>=0.7.6
python-dotenv>=1.0.0
